% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlparse-package.R
\docType{package}
\name{curlparse}
\alias{curlparse}
\alias{curlparse-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/curlparse}
\item BugReports: \url{https://gitlab.com/hrbrmstr/curlparse/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
