% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-valid-url.R
\name{is_valid_url}
\alias{is_valid_url}
\title{Test if a URL is a valid URL}
\usage{
is_valid_url(urls)
}
\arguments{
\item{urls}{character vector of URLs}
}
\value{
logical vector of test results
}
\description{
Test if a URL is a valid URL
}
\note{
Uses Diego Perini's "perfect" URL regex: \url{https://mathiasbynens.be/demo/url-regex}
}
