% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_curl}
\alias{parse_curl}
\title{Parse a character vector of URLs into component parts}
\usage{
parse_curl(urls)
}
\arguments{
\item{urls}{character vector of URLs}
}
\value{
data frame (tibble)
}
\description{
Parse a character vector of URLs into component parts
}
