% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/Z-domain.R, R/domain.R
\name{scheme}
\alias{scheme}
\alias{user}
\alias{password}
\alias{host}
\alias{port}
\alias{path}
\alias{url_options}
\alias{query}
\alias{fragment}
\alias{domain}
\title{Extract member components from a URL string}
\usage{
scheme(urls)

user(urls)

password(urls)

host(urls)

port(urls)

path(urls)

url_options(urls)

query(urls)

fragment(urls)

domain(urls)

domain(urls)
}
\arguments{
\item{urls}{character vector of URLs}
}
\value{
character vector of the extracted URL component
}
\description{
Extract member components from a URL string
}
