% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{url_parse}
\alias{url_parse}
\title{Parse a character vector of URLs into component parts (\code{urltools} compatibility function)}
\usage{
url_parse(urls)
}
\arguments{
\item{urls}{character vector of URLs}
}
\value{
data frame (not a tibble)
}
\description{
Parse a character vector of URLs into component parts (\code{urltools} compatibility function)
}
