% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlparse-package.R
\docType{package}
\name{curlparse}
\alias{curlparse}
\alias{curlparse-package}
\title{Parse 'URLs' with 'libcurl'}
\description{
As of version 7.62.0 'libcurl' has exposed its 'URL' parser. Tools are
provided to parse 'URLs' using this new parser feature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/curlparse}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/curlparse/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
